dates
I found out how to for loop in intervals of eight, but the for loop ends on the 40th item without displaying it. In any case, it gives the current weather in the main table, so it technically constitutes as a five day forecast, and actually I dont think I can progress further than that without paying money.